requirements.txt
crewai
python-dotenv
pyyaml
litellm
fpdf2
markdown2
PyGithub
gitpython
